separated channel components and fixed styling
