+++ react-app/src/components/ChannelFeed/ChannelFeed.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { useDispatch, useSelector } from "react-redux";
 
-import ChannelMessages from "./ChannelMessages";
+import ChannelMessages from "../ChannelMessages";
 
 import styles from "./ChannelFeed.module.css";
 // className={styles. }
