+++ react-app/src/components/ChannelFeed/index.js
@@ -1,0 +1,3 @@
+import ChannelFeed from "./ChannelFeed";
+
+export default ChannelFeed;
