+++ react-app/src/components/ChannelMessages/index.js
@@ -1,0 +1,3 @@
+import ChannelMessages from "./ChannelMessages";
+
+export default ChannelMessages;
